+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # CHANGELOG
 ## [Unreleased]
 
+## [2.0.0](../../compare/1.0.0..2.0.0) - XXXX-XX-XX
+### Added
+- `getLength()` method
+- `[Symbol.iterator]` method to iterate through an enum in `for..of` loops
+
+### Changed
+- Constructor is overloadable now
+
+### Removed
+- Generic properties signature
+- `Enum.from<T>(properties)` method
+
 ## [1.0.0](../../tree/1.0.0) - 2021-10-06
 Release